+++ HashTable.h
@@ -24,7 +24,7 @@ public:
     void insert(int key, double value);
     void remove(int key);
 
-    float operator[](int key);
+    double operator[](int key);
 
     float getCurrLoadFactor();
     float getMaxLoadFactor();
